chore: Moves Radio to the components folder
