+++ superset-frontend/spec/javascripts/explore/components/SaveModal_spec.jsx
@@ -25,7 +25,7 @@ import { Provider } from 'react-redux';
 import { shallow } from 'enzyme';
 import { styledMount as mount } from 'spec/helpers/theming';
 import { FormControl } from 'react-bootstrap';
-import { Radio } from 'src/common/components/Radio';
+import { Radio } from 'src/components/Radio';
 import Button from 'src/components/Button';
 import sinon from 'sinon';
 import fetchMock from 'fetch-mock';
