+++ superset-frontend/spec/javascripts/views/CRUD/alert/AlertReportModal_spec.jsx
@@ -25,8 +25,8 @@ import { act } from 'react-dom/test-utils';
 import AlertReportModal from 'src/views/CRUD/alert/AlertReportModal';
 import Modal from 'src/components/Modal';
 import { AsyncSelect, NativeGraySelect as Select } from 'src/components/Select';
-import { Radio } from 'src/common/components/Radio';
 import { Switch } from 'src/components/Switch';
+import { Radio } from 'src/components/Radio';
 import TextAreaControl from 'src/explore/components/controls/TextAreaControl';
 import waitForComponentToPaint from 'spec/helpers/waitForComponentToPaint';
 import { styledMount as mount } from 'spec/helpers/theming';
