+++ superset-frontend/src/explore/components/SaveModal.tsx
@@ -22,8 +22,8 @@ import { FormControl, FormGroup } from 'react-bootstrap';
 import Alert from 'src/components/Alert';
 import { JsonObject, t, styled } from '@superset-ui/core';
 import ReactMarkdown from 'react-markdown';
-import { Radio } from 'src/common/components/Radio';
 import Modal from 'src/components/Modal';
+import { Radio } from 'src/components/Radio';
 import Button from 'src/components/Button';
 import FormLabel from 'src/components/FormLabel';
 import { CreatableSelect } from 'src/components/Select';
