+++ superset-frontend/src/views/CRUD/alert/components/AlertReportCronScheduler.tsx
@@ -20,7 +20,7 @@ import React, { useState, useCallback, useRef, FunctionComponent } from 'react';
 import { t, useTheme } from '@superset-ui/core';
 
 import { Input, AntdInput } from 'src/common/components';
-import { Radio } from 'src/common/components/Radio';
+import { Radio } from 'src/components/Radio';
 import { CronPicker, CronError } from 'src/components/CronPicker';
 import { StyledInputContainer } from 'src/views/CRUD/alert/AlertReportModal';
